% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ask_ok_to_store}
\alias{ask_ok_to_store}
\title{Ask the user if it is OK to store files on their computer}
\usage{
ask_ok_to_store(
  message = "The onetime package requests to store files in '\%s'.",
  confirm_prompt = "Is this OK? [Yn] ",
  confirm_answers = c("Y", "y", "Yes", "yes", "YES"),
  default_answer = "Y"
)
}
\arguments{
\item{message}{Message to display to the user.}

\item{confirm_prompt}{Character string. Question to prompt the user to hide
the message in future.}

\item{confirm_answers}{Character vector. Answers which will cause
the message to be hidden in future.}

\item{default_answer}{Character string. Default answer if user
simply presses return.}
}
\value{
\code{TRUE} if the user gives permission (or has done so previously).
Does nothing and returns \code{FALSE} if not interactive. \code{FALSE} if the user
does not give permission.
}
\description{
Ask the user if it is OK to store files on their computer
}
\examples{
\dontrun{
  ask_ok_to_store()
}
}
