% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{check_ok_to_store}
\alias{check_ok_to_store}
\title{Check if the package has permission to store files on the user's computer}
\usage{
check_ok_to_store(ask = FALSE, ...)
}
\arguments{
\item{ask}{\code{TRUE} to call \code{\link[=ask_ok_to_store]{ask_ok_to_store()}} if necessary.}

\item{...}{Passed to \code{\link[=ask_ok_to_store]{ask_ok_to_store()}} if \code{ask} is \code{TRUE}.}
}
\value{
\code{TRUE} if:
\itemize{
\item We already have permission;
\item \code{ask} is \code{TRUE}, we are in an interactive session and the user
gives us permission;
\item \code{options("onetime.dir")} is set to a non-\code{NULL} value
}

Otherwise \code{FALSE}.
}
\description{
Check if the package has permission to store files on the user's computer
}
\examples{
\dontrun{
check_ok_to_store()
}
}
