% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang.R
\name{onetime-rlang}
\alias{onetime-rlang}
\alias{onetime_rlang_warn}
\alias{onetime_rlang_inform}
\title{Print a warning or message only once using \code{rlang} functions}
\usage{
onetime_rlang_warn(
  ...,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL
)

onetime_rlang_inform(
  ...,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL
)
}
\arguments{
\item{...}{Passed to \code{\link[rlang:abort]{rlang::warn()}} or \code{\link[rlang:abort]{rlang::inform()}}}

\item{id}{Unique ID string. By default, name of the calling package.}

\item{path}{Directory to store lockfiles.}

\item{expiry}{\code{\link[=difftime]{difftime()}} or e.g. \code{\link[lubridate:duration]{lubridate::duration()}} object.
After this length of time, code will be run again.}
}
\value{
\code{TRUE} if the message/warning was shown, \code{FALSE} otherwise.
}
\description{
If rlang is not installed, these fall back to \code{\link[=onetime_warning]{onetime_warning()}} and
\code{\link[=onetime_message]{onetime_message()}}, adding a note about rlang.
}
\examples{
\dontrun{
id <- sample(10000L, 1)
for (n in 1:3) {
  onetime_rlang_warn(c("rlang-style warning", i = "Extra info"), id = id)
}

onetime_reset(id = id)
}
}
