% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_been_done}
\alias{onetime_been_done}
\title{Check if a onetime call has already been made}
\usage{
onetime_been_done(id = calling_package(), path = default_lockfile_dir())
}
\arguments{
\item{id}{Unique ID string. By default, name of the calling package.}

\item{path}{Directory to store lockfiles.}
}
\value{
\code{TRUE} if the lockfile recording the
onetime call exists, \code{FALSE} otherwise. Note that
this doesn't take \code{expiry} into account.
}
\description{
Check if a onetime call has already been made
}
\examples{
\dontrun{
id <- sample(10000L, 1)
onetime_been_done(id = id)
onetime_do(cat("Creating an ID"),  id = id)
onetime_been_done(id = id)

onetime_reset(id = id)
}
}
