% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_do}
\alias{onetime_do}
\title{Run code only once}
\usage{
onetime_do(expr, id = calling_package(), path = lockfile_dir())
}
\arguments{
\item{expr}{The code to evaluate}

\item{id}{Unique ID. By default, name of the calling package.}

\item{path}{Path to store lockfiles. By default,
`system.file("lockfiles", package = "onetime")`.}
}
\value{
The value of `expr`, or `NULL` if called the second time.
}
\description{
Run code only once
}
\examples{
\dontrun{
id <- sample(10000L, 1)
for (n in 1:3) {
  onetime_do(print("printed once"), id = id)
}
}
}
