% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_message_confirm}
\alias{onetime_message_confirm}
\title{Print a message, and ask if it should be printed again}
\usage{
onetime_message_confirm(
  message,
  id = calling_package(),
  path = default_lockfile_dir(),
  confirm_prompt = "Show this message again? [yN]",
  hide_answers = c("N", "n", "No", "no"),
  default_answer = "N"
)
}
\arguments{
\item{message}{Message to print}

\item{id}{Unique ID string. By default, name of the calling package.}

\item{path}{Directory to store lockfiles.}

\item{confirm_prompt}{Question to prompt the user to hide the
message in future}

\item{hide_answers}{Character vector. Answers which will cause
the message to be hidden in future.}

\item{default_answer}{Character string. Default answer if user
simply presses return.}
}
\value{
\code{NULL} if the message was not shown (shown already or non-interactive
session). \code{TRUE} if the user asked to show the message again.
\code{FALSE} if the user asked to hide the message.
}
\description{
This uses \code{\link[=readline]{readline()}} to ask the user if the message should
be shown again in future. In a non-interactive session, it does
nothing.
}
\examples{
\dontrun{
id <- sample(10000L, 1)
onetime_message_confirm("A message to show one or more times", id = id)
}
}
