% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{set_ok_to_store}
\alias{set_ok_to_store}
\title{Grant or revoke permission to store lockfiles on the user's computer}
\usage{
set_ok_to_store(ok = TRUE)
}
\arguments{
\item{ok}{\code{TRUE} to grant permission to store lockfiles, \code{FALSE} to revoke
it and unset \code{options("onetime.dir")}#'}
}
\value{
\code{TRUE} if the operation succeeded
}
\description{
Grant or revoke permission to store lockfiles on the user's computer
}
\examples{
\dontrun{
set_ok_to_store()
}
}
