% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\docType{package}
\name{onetime}
\alias{onetime}
\title{Run code only once}
\description{
Run code only once
}
\details{
\code{onetime} provides convenience functions to run R code only once on a
given computer. For example, you can show a startup message only the
first time a package is loaded.
\subsection{Example}{\if{html}{\out{<div class="r">}}\preformatted{library(onetime)
ids  <- replicate(3, paste0("onetime-readme-", sample(1e9, 1)))


for (i in 1:5) \{
  cat("Loop ", i, " of 5\n")
  onetime_do(cat("You will only see this once.\n"), id = ids[1])
  onetime_warning("This warning will only be shown once.", 
      id = ids[2])
  onetime_message("This message will only be shown once.", id = ids[3])
\}
}\if{html}{\out{</div>}}\preformatted{## Loop  1  of 5
## You will only see this once.

## Warning in eval(expr, p): This warning will only be shown once.

## This message will only be shown once.

## Loop  2  of 5
## Loop  3  of 5
## Loop  4  of 5
## Loop  5  of 5
}
}
}
