% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_only}
\alias{onetime_only}
\title{Wrap a function to be called only once}
\usage{
onetime_only(.f, id = calling_package(), path = default_lockfile_dir())
}
\arguments{
\item{.f}{A function}

\item{id}{Unique ID. By default, name of the calling package.}

\item{path}{Directory to store lockfiles.}
}
\value{
A wrapped function.
}
\description{
This takes a function and returns the same function wrapped by \code{\link[=onetime_do]{onetime_do()}}.
Use it for code which should run only once, but which may be called from
multiple locations. This frees you from having to use the same \code{id} multiple
times.
}
\examples{
\dontrun{
id <- sample(10000L, 1)

cat_once <- onetime_only(cat, id = id)
cat_once("Purrr!")
cat_once("Miaow!")

onetime_reset(id)
}
}
