% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_reset}
\alias{onetime_reset}
\title{Reset a onetime call by ID}
\usage{
onetime_reset(id, path = lockfile_dir(calling_package()))
}
\arguments{
\item{id}{Unique ID. By default, name of the calling package.}

\item{path}{Directory to store lockfiles. By default, a
subdirectory based on the calling package, within a
directory \code{"onetime-lockfiles"}, which is itself within
\code{\link[rappdirs:user_config_dir]{rappdirs::user_config_dir()}}.}
}
\value{
The result of \code{file.remove()}, invisibly.
}
\description{
Reset a onetime call by ID
}
\examples{
\dontrun{
id <- sample(10000L, 1)
onetime_do(print("will be shown"),  id = id)
onetime_do(print("won't be shown"), id = id)
onetime_reset(id = id)
onetime_do(print("will be shown"),  id = id)
}
}
