% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_warning}
\alias{onetime_warning}
\alias{onetime_message}
\alias{onetime_startup_message}
\title{Print a warning or message only once}
\usage{
onetime_warning(
  ...,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL
)

onetime_message(
  ...,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL
)

onetime_startup_message(
  ...,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL
)
}
\arguments{
\item{...}{Passed to \code{\link[=warning]{warning()}}, \code{\link[=message]{message()}} or \code{\link[=packageStartupMessage]{packageStartupMessage()}}.}

\item{id}{Unique ID string. By default, name of the calling package.}

\item{path}{Directory to store lockfiles.}

\item{expiry}{\code{\link[=difftime]{difftime()}} or e.g. \code{\link[lubridate:duration]{lubridate::duration()}} object.
After this length of time, code will be run again.}
}
\value{
\code{TRUE} if the message/warning was shown, \code{FALSE} otherwise.
}
\description{
These functions use \code{\link[=onetime_do]{onetime_do()}} to print a warning or message just
once.
}
\examples{
\dontrun{
id <- sample(10000L, 1)
for (n in 1:3) {
  onetime_warning("will be shown once", id = id)
}
}
}
\seealso{
\code{\link[=onetime_do]{onetime_do()}}
}
