% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_warning}
\alias{onetime_warning}
\alias{onetime_message}
\title{Print a warning or message only once}
\usage{
onetime_warning(
  ...,
  id = calling_package(),
  path = lockfile_dir(calling_package())
)

onetime_message(
  ...,
  id = calling_package(),
  path = lockfile_dir(calling_package())
)
}
\arguments{
\item{...}{Passed to \code{\link[=warning]{warning()}} or \code{\link[=message]{message()}}.}

\item{id}{Unique ID. By default, name of the calling package.}

\item{path}{Directory to store lockfiles. By default, a
subdirectory based on the calling package, within a
directory \code{"onetime-lockfiles"}, which is itself within
\code{\link[rappdirs:user_config_dir]{rappdirs::user_config_dir()}}.}
}
\value{
The return value of \code{warning()/message()}, or \code{NULL} if
called a second time.
}
\description{
These functions use \code{\link[=onetime_do]{onetime_do()}} to print a warning or message just
once.
}
\examples{
\dontrun{
id <- sample(10000L, 1)
for (n in 1:3) {
  onetime_warning("will be shown once", id = id)
}
}
}
