% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{onetime_warning}
\alias{onetime_warning}
\alias{onetime_message}
\title{Print a warning or message only once}
\usage{
onetime_warning(..., id = calling_package(), path = lockfile_dir())

onetime_message(..., id = calling_package(), path = lockfile_dir())
}
\arguments{
\item{...}{Passed to [warning()] or [message()].}

\item{id}{Unique ID. By default, name of the calling package.}

\item{path}{Path to store lockfiles. By default,
`system.file("lockfiles", package = "onetime")`.}
}
\value{
The return value of `warning()/message()`, or `NULL` if
  called a second time.
}
\description{
Print a warning or message only once
}
\examples{
\dontrun{
id <- sample(10000L, 1)
for (n in 1:3) {
  onetime_warning("will be shown once", id = id)
}
}
}
