% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{set_ok_to_store}
\alias{set_ok_to_store}
\title{Grant or revoke permission to store lockfiles on the user's computer}
\usage{
set_ok_to_store(ok = TRUE)
}
\arguments{
\item{ok}{\code{TRUE} to grant permission to store lockfiles, \code{FALSE} to revoke
it and unset \code{options("onetime.dir")}.}
}
\value{
Invisible \code{NULL}.
}
\description{
End users may use this from the command line. Package authors should \emph{only}
call it if they cannot ask for permission interactively using
\code{check_ok_to_store(ask = TRUE)}.
}
\examples{
\dontrun{
set_ok_to_store()
}
}
